requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
python-dotenv
pytest
flake8